Update Lazyload for Dropdown
Fix bugs to ensure that lazyload works reliably when multiple dropdowns are present on the page.

Lazyloading the dropdown saves a remarkable amount of time, ensuring that the list of options is constructed on demand. This means that the logic that runs to transfer the data to the dropdown is delayed until the user actually clicks on it.